pyproject.toml
[project]
name = "python-mcp-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.5",
    "debugpy>=1.8.0",
    "langchain-core>=0.3.0",
    "mcp>=1.3.0",
    "azure-identity>=1.25.1",
    "langchain==1.0.0a5",
    "langchain-openai>=1.0.1",
    "langchain-mcp-adapters>=0.1.11",
    "azure-ai-agents>=1.1.0",
    "agent-framework>=1.0.0b251016",
]
